pyproject.toml
[project]
name = "contextctl"
version = "0.1.0"
description = "ContextCtl is a repo-aware prompt and rule manager. It pulls the right context for any codebase from a central store, so teams can use consistent, up-to-date prompts with any AI tool."
readme = "README.md"
authors = [{ name = "Darec Stowell", email = "darec.specmaker@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contextctl"]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.2.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.5",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 119
target-version = "py313"
exclude = [".git", ".ruff_cache", ".venv", ".vscode"]
src = ["src"]

[tool.ruff.lint]
preview = true
select = [
    "ANN", # type annotation
    "B",   # flake8-bugbear
    "D",   # pydocstyle
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "PTH", # use `pathlib.Path` instead of `os.path`
    "RUF", # ruff specific rules
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]
ignore = [
    "ANN401", # Checks that function arguments are annotated with a more specific type than Any.
    "B007",   # Unused loop variable
    "B008",   # Function call in default argument
    "B905",   # `zip()` without `strict=True`
    "COM812", # Missing trailing comma
    "COM819", # Trailing comma prohibited
    "D1",     # Missing docstring in public module, class, function, or method
    "D203",   # One blank line required before class docstring (ignored because D211 is prioritized in pydocstyle convention 'google')
    "D205",   # 1 blank line required between summary line and description
    "D212",   # Multi-line docstring summary should start at the first line
    "D213",   # Multi-line docstring summary should start at the second line
    "D400",   # First line of docstring should end with a period
    "D415",   # First line of docstring should end with punctuation (period, question mark, or exclamation point)
    "E114",   # Indentation is not a multiple of four (comment)
    "G004",   # Logging statement uses f-string
    "ISC001", # Implicit string concatenation on a single line
    "ISC002", # Implicit string concatenation on multiple lines
    "PTH123", # `open()` should be replaced by `Path.open()`
    "Q000",   # Single quotes found when double quotes are preferred
    "Q001",   # Single quotes found when double quotes are preferred for multi-line strings
    "Q002",   # Single quotes found when double quotes are preferred for docstrings
    "RUF002", # Checks for ambiguous Unicode characters in docstrings.
    "RUF003", # Checks for ambiguous Unicode characters in comments.
    "SIM105", # try-except-pass blocks that can be replaced with the contextlib.suppress context manager.
    "SIM108", # Use ternary operator instead of `if`-`else`-block
    "SIM116", # Use a dictionary instead of consecutive `if` statements
]
unfixable = [
    "F401", # unused import
    "F841", # unused variable
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
plugins = []
mypy_path = ["src"]

[tool.pytest.ini_options]
addopts = "--strict-markers --disable-warnings"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "error::DeprecationWarning",
]
